+++ flask_sqlalchemy_unchained.py
@@ -2,8 +2,8 @@ from flask_sqlalchemy import SQLAlchemy as _SQLAlchemy, BaseQuery as _Query
 from sqlalchemy.ext.associationproxy import association_proxy
 from sqlalchemy.ext.declarative import declared_attr
 from sqlalchemy.ext.hybrid import hybrid_method, hybrid_property
-from sqlalchemy_unchained import (BaseModel, DeclarativeMeta, _ModelRegistry,
-                                  QueryMixin, declarative_base, foreign_key)
+from sqlalchemy_unchained import (BaseModel, DeclarativeMeta, QueryMixin,
+                                  declarative_base, foreign_key)
 
 
 class BaseQuery(QueryMixin, _Query):
@@ -16,8 +16,6 @@ class SQLAlchemyUnchained(_SQLAlchemy):
         super().__init__(app=app, use_native_unicode=use_native_unicode,
                          session_options=session_options, metadata=metadata,
                          query_class=query_class, model_class=model_class)
-        _ModelRegistry().register_base_model_class(self.Model)
-
         self.association_proxy = association_proxy
         self.declared_attr = declared_attr
         self.foreign_key = foreign_key